lily.py: LilySchema, bug fix,  box finishing at last offset of a part
getElementByOffset() returns an empty array when the start offset is the last offset of the part.
This offset does not match any music21 object (measure for example).
So in this case the last measure of the part is selected to place the end of the box.
The offsetEnd arg of getElementByOffset() has been commented because I thought it was not useful...